@@ -1,13 +1,13 @@
 package com.qingmei2.samplepaging.adapter
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
 import com.qingmei2.samplepaging.R
 import com.qingmei2.samplepaging.db.Student
 
-class StudentViewHolder(parent: ViewGroup) : RecyclerView.ViewHolder(
+class StudentViewHolder(parent: ViewGroup) : androidx.recyclerview.widget.RecyclerView.ViewHolder(
         LayoutInflater.from(parent.context).inflate(R.layout.student_item, parent, false)) {
 
     private val nameView = itemView.findViewById<TextView>(R.id.name)
