@@ -1,7 +1,7 @@
 package com.qingmei2.samplepaging.db
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity
 data class Student(@PrimaryKey(autoGenerate = true) val id: Int,
